requirements.txt
appdirs
attrs
black
boto3
botocore
Click
coverage
docutils
jmespath
mock
nose
pathspec
policyuniverse
python-dateutil
regex
s3transfer
six
toml
typed-ast
urllib3
pytz
pyyaml
jira
